Bundle states and block archiver into chain
