+++ packages/lodestar/src/chain/chain.ts
@@ -30,6 +30,7 @@ import {AttestationPool, SyncCommitteeMessagePool, SyncContributionAndProofPool}
 import {ForkDigestContext, IForkDigestContext} from "../util/forkDigestContext";
 import {LightClientIniter} from "./lightClient";
 import {AggregatedAttestationPool} from "./opPools/aggregatedAttestationPool";
+import {Archiver} from "./archiver";
 
 export interface IBeaconChainModules {
   config: IBeaconConfig;
@@ -54,6 +55,7 @@ export class BeaconChain implements IBeaconChain {
   forkDigestContext: IForkDigestContext;
   lightclientUpdater: LightClientUpdater;
   lightClientIniter: LightClientIniter;
+  archiver: Archiver;
 
   // Ops pool
   readonly attestationPool = new AttestationPool();
@@ -141,6 +143,7 @@ export class BeaconChain implements IBeaconChain {
 
     this.lightclientUpdater = new LightClientUpdater(this.db);
     this.lightClientIniter = new LightClientIniter({config: this.config, forkChoice, db: this.db, stateCache});
+    this.archiver = new Archiver(db, this, logger, signal);
 
     handleChainEvents.bind(this)(this.abortController.signal);
   }
@@ -149,6 +152,10 @@ export class BeaconChain implements IBeaconChain {
     this.abortController.abort();
     this.stateCache.clear();
     this.checkpointStateCache.clear();
+  }
+
+  async persistToDisk(): Promise<void> {
+    await this.archiver.persistToDisk();
   }
 
   getGenesisTime(): Number64 {
