--- packages/lodestar/src/tasks/interface.ts
@@ -1,7 +1,0 @@
-/**
- * @module chores
- */
-
-export interface ITask {
-  run(): Promise<void>;
-}
